reorganize event definitions
