+++ SpiffWorkflow/bpmn/parser/TaskParser.py
@@ -26,7 +26,7 @@ from SpiffWorkflow.bpmn.specs.defaults import (
     ParallelMultiInstanceTask
 )
 from SpiffWorkflow.bpmn.specs.control import _BoundaryEventParent
-from SpiffWorkflow.bpmn.specs.event_definitions import CancelEventDefinition
+from SpiffWorkflow.bpmn.specs.event_definitions.simple import CancelEventDefinition
 from SpiffWorkflow.bpmn.specs.data_spec import TaskDataReference
 
 from .util import one
